fix(opencode): harden shared install contract checks
