+++ docker/agent/Dockerfile
@@ -241,11 +241,17 @@ RUN if [ -z "${KILOCODE_INSTALL_COMMAND}" ]; then \
 # OpenCode CLI: install contract is owned by agent-harness.
 # OPENCODE_INSTALL_COMMAND is extracted from agent-harness at build time by
 # build-agent-image.sh / agent-image.yml — Paid no longer pins this version.
+# Supply-chain hardening: enforce --ignore-scripts regardless of what the
+# contract returns, matching the policy applied to every other npm install above.
 ARG OPENCODE_INSTALL_COMMAND
 RUN if [ -z "${OPENCODE_INSTALL_COMMAND}" ]; then \
         echo "ERROR: OPENCODE_INSTALL_COMMAND not set; aborting build" >&2; \
         exit 1; \
     fi \
+    && case "${OPENCODE_INSTALL_COMMAND}" in \
+        *--ignore-scripts*) ;; \
+        *) echo "ERROR: OPENCODE_INSTALL_COMMAND must include --ignore-scripts for supply-chain safety" >&2; exit 1 ;; \
+    esac \
     && echo "Installing OpenCode via agent-harness contract: ${OPENCODE_INSTALL_COMMAND}" \
     && eval "${OPENCODE_INSTALL_COMMAND}"
